LPS-105599 Move all floating toolbar components
Floating toolbar related components are now in their
own directory
