+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -5,7 +5,7 @@
 @import './DragDrop';
 @import './DragPreview';
 @import './DropZone';
-@import './FloatingToolbar';
+@import './floating-toolbar/FloatingToolbar';
 @import './MasterLayout';
 @import './layout-data-items/ColumnOverlayGrid';
 @import './layout-data-items/FragmentsEditorShim';
