+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/FloatingToolbar.js
@@ -26,9 +26,9 @@ import React, {
 } from 'react';
 import {createPortal} from 'react-dom';
 
-import debounceRAF from '../../core/debounceRAF';
-import {FLOATING_TOOLBAR_CONFIGURATIONS} from '../config/constants/floatingToolbarConfigurations';
-import {useIsActive} from './Controls';
+import debounceRAF from '../../../core/debounceRAF';
+import {FLOATING_TOOLBAR_CONFIGURATIONS} from '../../config/constants/floatingToolbarConfigurations';
+import {useIsActive} from '../Controls';
 
 export default function FloatingToolbar({
 	buttons,
