+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/FloatingToolbar.scss
@@ -1,4 +1,4 @@
-@import 'variables';
+@import '../variables';
 
 @keyframes FloatingToolbarPanelFadeIn {
 	to {
