+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/FragmentConfigurationPanel.js
@@ -16,12 +16,12 @@ import ClayButton from '@clayui/button';
 import ClayIcon from '@clayui/icon';
 import React, {useContext} from 'react';
 
-import {FREEMARKER_FRAGMENT_ENTRY_PROCESSOR} from '../../../js/utils/constants';
-import {FRAGMENT_CONFIGURATION_FIELD_TYPES} from '../config/constants/fragmentConfigurationFieldTypes';
-import {ConfigContext} from '../config/index';
-import selectPrefixedSegmentsExperienceId from '../selectors/selectPrefixedSegmentsExperienceId';
-import {useSelector, useDispatch} from '../store/index';
-import updateFragmentConfiguration from '../thunks/updateFragmentConfiguration';
+import {FREEMARKER_FRAGMENT_ENTRY_PROCESSOR} from '../../../../js/utils/constants';
+import {FRAGMENT_CONFIGURATION_FIELD_TYPES} from '../../config/constants/fragmentConfigurationFieldTypes';
+import {ConfigContext} from '../../config/index';
+import selectPrefixedSegmentsExperienceId from '../../selectors/selectPrefixedSegmentsExperienceId';
+import {useSelector, useDispatch} from '../../store/index';
+import updateFragmentConfiguration from '../../thunks/updateFragmentConfiguration';
 
 const FieldSet = ({configurationValues, fields, label, onValueSelect}) => {
 	return (
