+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/FragmentWithControls.js
@@ -35,8 +35,8 @@ import selectShowLayoutItemTopper from '../../selectors/selectShowLayoutItemTopp
 import {useSelector, useDispatch} from '../../store/index';
 import duplicateItem from '../../thunks/duplicateItem';
 import {useSelectItem} from '../Controls';
-import FloatingToolbar from '../FloatingToolbar';
 import Topper from '../Topper';
+import FloatingToolbar from '../floating-toolbar/FloatingToolbar';
 import FragmentContent from './FragmentContent';
 
 const FragmentWithControls = React.forwardRef(({item, layoutData}, ref) => {
