+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/components/RowConfigurationPanel.test.js
@@ -16,7 +16,7 @@ import '@testing-library/jest-dom/extend-expect';
 import {cleanup, fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
-import {RowConfigurationPanel} from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/RowConfigurationPanel';
+import {RowConfigurationPanel} from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/RowConfigurationPanel';
 import {ConfigContext} from '../../../../src/main/resources/META-INF/resources/page_editor/app/config';
 import {StoreAPIContextProvider} from '../../../../src/main/resources/META-INF/resources/page_editor/app/store';
 import updateItemConfig from '../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
